+++ packages/mg-components/src/components/molecules/inputs/mg-input-checkbox/tests/mg-input-checkbox.e2e.ts
@@ -484,4 +484,40 @@ test.describe('mg-input-checkbox', () => {
       await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
     });
   });
+  test('Should udpate error with displayError() after value update with props', async ({page}) => {
+    const componentArgs = {
+      ...baseArgs,
+      required: true,
+      value: baseArgs.value.map((val, i) => ({...val, value: i === 0}))
+    }
+    const html = createHTML(componentArgs);
+
+    await page.setContent(html);
+    await page.addScriptTag({ content: renderProperties(componentArgs, `[identifier="${componentArgs.identifier}"]`) });
+
+    await page.locator('mg-input-checkbox.hydrated').waitFor();
+    await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
+
+    // Programaticaly remove value to display required error message
+    await page.locator('mg-input-checkbox').evaluate(async (elm: HTMLMgInputTextElement, value) => {
+      elm.value = value;
+      await elm.displayError()
+    }, baseArgs.value);
+
+    // Check state 
+    // - without any selected value
+    // - with error message
+    await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
+
+    // Enter a new value from JS and remove required error
+    await page.locator('mg-input-checkbox').evaluate(async (elm: HTMLMgInputTextElement, value) => {
+      elm.value = value.map((val, i) => ({...val, value: i === 2}));
+      await elm.displayError()
+    }, baseArgs.value);
+
+    // Check state 
+    // - with selected value
+    // - without error message
+    await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
+  })
 });