+++ packages/mg-components/src/components/molecules/inputs/mg-input-textarea/tests/mg-input-textarea.spec.tsx
@@ -5,7 +5,7 @@ import messages from '../../../../../locales/en/messages.json';
 import { MgInput } from '../../mg-input/mg-input';
 import { MgInputTitle } from '../../../../atoms/internals/mg-input-title/mg-input-title';
 import { tooltipPositions } from '../../mg-input/mg-input.conf';
-import { toString } from '@mgdis/stencil-helpers';
+import { setUpRequestAnimationFrameMock, toString } from '@mgdis/stencil-helpers';
 
 const getPage = args => {
   const page = newSpecPage({
@@ -14,6 +14,7 @@ const getPage = args => {
   });
 
   jest.runAllTimers();
+  setUpRequestAnimationFrameMock(jest.runOnlyPendingTimers)
 
   return page;
 };