+++ packages/mg-components/src/components/molecules/internals/mg-item-more/tests/mg-item-more.e2e.ts
@@ -1,7 +1,6 @@
 import { expect } from '@playwright/test';
 import { renderAttributes } from '@mgdis/playwright-helpers';
 import { test } from '../../../../../utils/playwright.fixture';
-import { OverflowBehaviorElements } from '../../../../../utils/behaviors.utils';
 import { Status } from '../../../menu/mg-menu-item/mg-menu-item.conf';
 import { Direction, type MenuSizeType, sizes } from '../../../menu/mg-menu/mg-menu.conf';
 
@@ -25,10 +24,10 @@ const getSubMenuSize = (size: MenuSizeType) => (size === 'medium' ? 'large' : 'm
 const createHTML = args => `
   <header class="menu-container menu-container--${args.direction}-small">
     <mg-menu ${renderAttributes({ label: 'menu', ...args })}>
-      <mg-menu-item status="active">
+      <mg-menu-item>
         <span slot="label">1 - head-1</span>
         <mg-menu ${renderAttributes({ label: 'sub-menu 1', direction: Direction.VERTICAL, size: getSubMenuSize(args?.size) })}>
-          <mg-menu-item><span slot="label">Batman begins</span></mg-menu-item>
+          <mg-menu-item status="active"><span slot="label">Batman begins</span></mg-menu-item>
         </mg-menu>
       </mg-menu-item>
       <mg-menu-item status="disabled"><span slot="label">1 - head-2 long</span></mg-menu-item>
@@ -84,12 +83,12 @@ test.describe('mg-item-more', () => {
 
         await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
 
-        await page.locator(`[${OverflowBehaviorElements.BASE_INDEX}="0"]`).evaluate((elm, status) => {
-          elm.setAttribute('status', status as string);
+        await page.getByText(`Batman begins`).first().evaluate((elm, status) => {
+          elm.closest('mg-menu-item').setAttribute('status', status as string);
         }, Status.VISIBLE);
 
-        await page.locator(`[${OverflowBehaviorElements.BASE_INDEX}="2"]`).evaluate((elm, status) => {
-          elm.setAttribute('status', status as string);
+        await page.getByText(`1 - head-4`).first().evaluate((elm, status) => {
+          elm.closest('mg-menu-item').setAttribute('status', status as string);
         }, Status.ACTIVE);
 
         await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();