+++ packages/mg-components/src/components/molecules/menu/mg-menu/mg-menu.tsx
@@ -1,4 +1,4 @@
-import { Component, h, Prop, State, Element, Watch, Host } from '@stencil/core';
+import { Component, h, Prop, Element, Watch, Host, Listen } from '@stencil/core';
 import { Direction, sizes } from './mg-menu.conf';
 import type { MenuSizeType, ItemMoreType, DirectionType } from './mg-menu.conf';
 import { toString } from '@mgdis/stencil-helpers';
@@ -17,9 +17,6 @@ export class MgMenu {
    ************/
 
   private readonly name = 'mg-menu';
-  private menuItems: HTMLMgMenuItemElement[] = [];
-  private focusedMenuItem = 0;
-  private itemMoreElement: HTMLMgItemMoreElement;
 
   /**************
    * Decorators *
@@ -50,6 +47,8 @@ export class MgMenu {
   validateDirection(newValue: MgMenu['direction']): void {
     if (![Direction.VERTICAL, Direction.HORIZONTAL].includes(newValue)) {
       throw new Error(`<${this.name}> prop "direction" must be one of: ${Direction.HORIZONTAL}, ${Direction.VERTICAL}. Passed value: ${toString(newValue)}.`);
+    } else {
+      this.updateMenuItems();
     }
   }
 
@@ -79,92 +78,83 @@ export class MgMenu {
   }
 
   /**
-   * is this menu a child menu. Used for conditional render.
+   * Define mg-menu-item listener when they are loaded
+   * @param event - mg-item-loaded "item-loaded" custom event
    */
-  @State() isChildMenu: boolean;
+  @Listen('item-loaded')
+  updateItemListener(event: CustomEvent & { target: HTMLMgMenuItemElement}): void {
+    const itemToWatch = event.target.nodeName === 'MG-ITEM-MORE' ? event.target.shadowRoot.querySelector('mg-menu-item') : event.target;
+
+    ['click', 'focus'].forEach(trigger => {
+      itemToWatch.addEventListener(trigger, ():void => {
+        const focusedItem: HTMLMgMenuItemElement = event.target.closest('mg-menu').querySelector('[data-has-focus]')
+        // exit process if new focused item in the current menu children
+        if(focusedItem === itemToWatch) return;
+
+        // reset expanded on previous active menu item
+        if(focusedItem) focusedItem.removeAttribute("data-has-focus");
+
+        // update focusedMenuItem with new value
+        itemToWatch.setAttribute('data-has-focus','true');
+      });
+    })
+  };
 
   /*************
    * Methods *
    *************/
 
   /**
-   * Close matching menu-item
-   * @param item - menu-item to close
-   * @param condition - addionnal condition
+   * Test if menu is a child menu
+   * @returns truthy if menu is a child menu
    */
-  private closeMenuItem = (item: HTMLMgMenuItemElement, condition: boolean): void => {
-    if (!this.isChildMenu && condition) {
-      item.expanded = false;
+  private isChildMenu = (): boolean => this.element.closest('mg-menu-item') !== null;
+
+  /**
+   * Get menu children items
+   * @returns menu children mg-menu-item
+   */
+  private getMenuItems = (menu = this.element): HTMLMgMenuItemElement[] => {
+    const menuItems = Array.from(menu.children).filter(child => child.nodeName === 'MG-MENU-ITEM') as HTMLMgMenuItemElement[];
+    const itemMoreMenuItem = menu.querySelector('mg-item-more')?.shadowRoot?.querySelector('mg-menu-item');
+    if(itemMoreMenuItem) {
+      menuItems.push(itemMoreMenuItem)
     }
-  };
+    return menuItems;
+  }
 
   /**
-   * get mg-item-more child mg-menu-item element
-   * @returns mg-menu-item element
+   * Update menu items
    */
-  private getItemMoreMenuItem = (): HTMLMgMenuItemElement => this.itemMoreElement?.shadowRoot?.querySelector('mg-menu-item');
+  private updateMenuItems = (): void => {
+    this.getMenuItems().forEach((item) => {
+      item.dataset.styleDirection = this.direction
+    })
 
-  /**
-   * Set item listend
-   * @param item - mg-menu-item element
-   * @param index - number
-   */
-  private setItemListener = (item: HTMLMgMenuItemElement, index: MgMenu['focusedMenuItem']): void => {
-    ['click', 'focus'].forEach(trigger => {
-      item.shadowRoot.querySelector('button,a').addEventListener(trigger, () => {
-        this.focusedMenuItem = index;
-        // reset expanded on previous active menu item
-        const itemMoreMenuItem = this.getItemMoreMenuItem();
-        (![null, undefined].includes(itemMoreMenuItem) ? [...this.menuItems, itemMoreMenuItem] : this.menuItems).forEach((item, index) => {
-          this.closeMenuItem(item, index !== this.focusedMenuItem);
-        });
-      });
-    });
-  };
-
-  /**
-   * Store menu-items on component init and add listeners
-   */
-  private initMenuItemsListeners = (): void => {
-    // add listeners on menu item and edit index
-    this.menuItems.forEach((item, index) => {
-      this.setItemListener(item, index);
-    });
-  };
-
-  /**
-   * Handle item-loaded event on mg-item-more element
-   */
-  private handleItemLoaded = (): void => {
-    this.setItemListener(this.getItemMoreMenuItem(), this.menuItems.length);
-  };
+    // render mg-item-more to manage OverflowBehavior
+    this.renderMgItemMore();
+  }
 
   /**
    * render mg-item-more
    */
   private renderMgItemMore = (): void => {
-    if (this.direction !== Direction.HORIZONTAL || this.isChildMenu || this.element.children.length <= 1) {
+    if (this.direction === Direction.VERTICAL || this.direction === Direction.HORIZONTAL && this.isChildMenu() || this.element.children.length <= 1) {
       return;
     }
 
     // Insert mg-item-more outside the mg-menu shadowdom
-    if (this.itemMoreElement === undefined) {
-      this.itemMoreElement = document.createElement('mg-item-more');
-      this.itemMoreElement.addEventListener('item-loaded', this.handleItemLoaded);
-      this.element.appendChild(this.itemMoreElement);
+    let itemMoreElement = this.element.querySelector('mg-item-more')
+    if (itemMoreElement === null) {
+      itemMoreElement = document.createElement('mg-item-more');
+      this.element.appendChild(itemMoreElement);
     }
 
     // update mg-item-more props
-    for (const attribute in this.itemmore) {
-      const newValue = this.itemmore[attribute];
-      // to improve rendering we use HTML attributes as much as possible
-      if (['string', 'number'].includes(typeof newValue)) {
-        this.itemMoreElement.setAttribute(attribute, newValue);
-      } else {
-        this.itemMoreElement[attribute] = newValue;
-      }
+    for (const propertie in this.itemmore) {
+      itemMoreElement[propertie] = this.itemmore[propertie];
     }
-    if (this.itemmore?.size === undefined) this.itemMoreElement.size = this.size;
+    if (this.itemmore?.size === undefined) itemMoreElement.size = this.size;
   };
 
   /*************
@@ -183,32 +173,29 @@ export class MgMenu {
 
   /**
    * Check if component slots configuration
-   * @returns timeout
    */
-  componentDidLoad(): ReturnType<typeof setTimeout> {
-    // update props and states after componentDidLoad hook
-    // return a promise to process action only in the FIRST render().
-    // https://stenciljs.com/docs/component-lifecycle#componentwillload
-    return setTimeout(() => {
-      // store all menu-items
-      this.menuItems = Array.from(this.element.children).filter(child => child.nodeName === 'MG-MENU-ITEM') as HTMLMgMenuItemElement[];
-      this.isChildMenu = this.element.closest('mg-menu-item') !== null;
+  componentDidLoad(): void {
+    // set menu-items
+    this.updateMenuItems();
 
-      // add mg-item-more to manage OverflowBehavior
-      // when all props, states and DOM are already rendered we can render the mg-item-more
-      this.renderMgItemMore();
-      // then use mutation observer on children to improve reactivity
-      new MutationObserver(() => {
-        this.renderMgItemMore();
-      }).observe(this.element, { childList: true });
-    });
-  }
-
-  /**
-   * Add listeners to items
-   */
-  componentDidRender(): void {
-    this.initMenuItemsListeners();
+    // add listener to close main menu when focus leave it
+    if(!this.isChildMenu() && this.direction === Direction.HORIZONTAL) {
+      document.addEventListener('focusin', (event: FocusEvent & { target: HTMLElement}):void => {
+        const isElementChild = (element: HTMLElement):boolean => {
+          if (element === this.element) return true;
+          else if(!element.parentElement) return false;
+          else return isElementChild(element.parentElement);
+        };
+  
+        if(!event.target.closest('mg-menu') || !isElementChild(event.target)) {
+          const focusedItem = this.element.querySelector('[data-has-focus]')
+          // reset expanded on previous active menu item
+          if(focusedItem) focusedItem.removeAttribute('data-has-focus');
+        }
+      })
+    }
+    // add mutation observer to improve global reactivity
+    new MutationObserver(this.updateMenuItems).observe(this.element, { childList: true });
   }
 
   /**