+++ packages/mg-components/src/components/molecules/menu/mg-menu/tests/mg-menu.spec.tsx
@@ -19,12 +19,12 @@ let id;
  */
 const setId = (hasId: boolean): string => (hasId ? `test-${id++}` : undefined);
 
-const getPage = async (args, options = { submenu: true, itemMore: false }) => {
+const getPage = async (args, options = { submenu: true, itemMore: false, siblingMenu: false }) => {
   const components: unknown[] = [MgMenu, MgMenuItem, MgPopover, MgPopoverContent];
   if (options.itemMore) components.push(MgItemMore);
   const page = await newSpecPage({
     components,
-    template: () => (
+    template: () => [
       <mg-menu {...args}>
         <mg-menu-item id={setId(args.hasId)} identifier="identifier-1">
           <span slot="label">batman</span>
@@ -58,8 +58,13 @@ const getPage = async (args, options = { submenu: true, itemMore: false }) => {
         <mg-menu-item href="#bane" id={setId(args.hasId)} identifier="identifier-3">
           <span slot="label">bane</span>
         </mg-menu-item>
+      </mg-menu>,
+      options.siblingMenu && <mg-menu label='sibling menu'>
+        <mg-menu-item identifier="identifier-1">
+          <span slot="label">sibling item</span>
+        </mg-menu-item>
       </mg-menu>
-    ),
+    ],
   });
 
   // flush mg-menu timeout and render
@@ -86,7 +91,7 @@ const getPage = async (args, options = { submenu: true, itemMore: false }) => {
 };
 
 describe('mg-menu', () => {
-  let fireMo;
+  let fireMo = [];
   beforeEach(() => {
     id = 1;
     jest.useFakeTimers({ legacyFakeTimers: true });
@@ -100,17 +105,20 @@ describe('mg-menu', () => {
     });
     setupMutationObserverMock({
       observe: function () {
-        fireMo = this.cb;
+        fireMo.push(this.cb);
       },
       disconnect: () => null,
       takeRecords: () => [],
     });
   });
 
-  afterEach(() => jest.clearAllTimers());
+  afterEach(() => {
+    jest.clearAllTimers();
+    fireMo = [];
+  });
 
   describe('render', () => {
-    test.each([{}, { direction: 'horizontal' }, { direction: 'vertical' }, { itemmore: { icon: { icon: 'user' } } }])('with args %s', async args => {
+    test.each([{}, { direction: Direction.HORIZONTAL }, { direction: Direction.VERTICAL }, { itemmore: { icon: { icon: 'user' } } }])('with args %s', async args => {
       const { root } = await getPage({ label: 'batman menu', ...args });
 
       expect(root).toMatchSnapshot();
@@ -120,7 +128,7 @@ describe('mg-menu', () => {
   describe('errors', () => {
     const baseProps = { label: 'batman menu' };
     test.each([
-      { props: { direction: 'horizontal' }, error: `<mg-menu> prop "label" is required. Passed value: undefined.` },
+      { props: { direction: Direction.HORIZONTAL }, error: `<mg-menu> prop "label" is required. Passed value: undefined.` },
       { props: { ...baseProps, direction: 'test' }, error: `<mg-menu> prop "direction" must be one of: horizontal, vertical. Passed value: test.` },
       { props: { ...baseProps, direction: Direction.VERTICAL, itemmore: { icon: 'user' } }, error: `<mg-menu> prop "itemmore" must be paired with direction horizontal.` },
       { props: { ...baseProps, size: 'batman' }, error: `<mg-menu> prop "size" must be one of: ${sizes.join(', ')}. Passed value: batman.` },
@@ -135,23 +143,45 @@ describe('mg-menu', () => {
     });
   });
 
-  describe.each(['horizontal', 'vertical'])('events', direction => {
-    test(`Should manage outside click, case direction ${direction}`, async () => {
-      const page = await getPage({ label: 'batman menu', direction });
+  describe.each([Direction.HORIZONTAL, Direction.VERTICAL])('events', direction => {
+    test.each(['click', 'focusin'])(`Should manage outside %s, case direction ${direction}`, async event => {
+      const page = await getPage({ label: 'batman menu', direction }, {submenu: true, siblingMenu: true, itemMore: false});
 
       const firstItem: HTMLMgMenuItemElement = page.root.querySelector('[title="batman"]').closest('mg-menu-item');
       expect(firstItem.expanded).toBe(false);
 
       firstItem.shadowRoot.querySelector('button').dispatchEvent(new MouseEvent('click', { bubbles: true }));
+      firstItem.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       await page.waitForChanges();
       jest.runOnlyPendingTimers();
 
       expect(firstItem.expanded).toBe(true);
 
-      document.dispatchEvent(new CustomEvent('click', { bubbles: true }));
+      if(event === 'focusin') {
+        const batmanChildItem: HTMLMgMenuItemElement = page.doc.querySelector('[title="batman begins"]').closest('mg-menu-item');
+        batmanChildItem.dispatchEvent(new MouseEvent(event, { bubbles: true }));
+        expect(firstItem.expanded).toBe(true);
+
+        const jokerItem: HTMLMgMenuItemElement = page.doc.querySelector('[title="joker"]').closest('mg-menu-item');
+        jokerItem.dispatchEvent(new MouseEvent(event, { bubbles: true }));
+        expect(firstItem.expanded).toBe(true);
+
+        const siblingMenuItem = page.doc.querySelector('[title="sibling item"]').closest('mg-menu-item')
+        siblingMenuItem.dispatchEvent(new MouseEvent(event, { bubbles: true }));
+        expect(firstItem.expanded).toBe(direction === Direction.VERTICAL);
+
+        if(direction === Direction.HORIZONTAL) {
+          firstItem.shadowRoot.querySelector('button').dispatchEvent(new MouseEvent('click', { bubbles: true }));
+          firstItem.dispatchEvent(new MouseEvent('click', { bubbles: true }));
+          await page.waitForChanges();
+          jest.runOnlyPendingTimers();
+        }
+      }
+
+      document.dispatchEvent(new MouseEvent(event, { bubbles: true }));
       await page.waitForChanges();
 
-      expect(firstItem.expanded).toBe(direction === 'vertical');
+      expect(firstItem.expanded).toBe(direction === Direction.VERTICAL);
     });
 
     test.each(['click', 'focus'])(`should manage sibling menu-item expanded props in ${direction} menu, case %s event`, async event => {
@@ -159,20 +189,20 @@ describe('mg-menu', () => {
       // open batman item
       const batmanItem: HTMLMgMenuItemElement = page.doc.querySelector('[title="batman"]').closest('mg-menu-item');
       expect(batmanItem.expanded).toBe(false);
-      if (direction === 'horizontal') {
+      if (direction === Direction.HORIZONTAL) {
         expect(batmanItem.shadowRoot.querySelector('mg-popover')).not.toBe(null);
       }
 
-      batmanItem.shadowRoot.querySelector('button').dispatchEvent(new CustomEvent('click', { bubbles: true }));
+      batmanItem.shadowRoot.querySelector('button').dispatchEvent(new MouseEvent('click', { bubbles: true }));
+      batmanItem.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       await page.waitForChanges();
-      jest.runOnlyPendingTimers();
 
       expect(batmanItem.expanded).toBe(true);
 
       // open batman first child item
       const batmanChildItem: HTMLMgMenuItemElement = page.doc.querySelector('[title="batman begins"]').closest('mg-menu-item');
-
-      batmanChildItem.shadowRoot.querySelector('button').dispatchEvent(new CustomEvent('click', { bubbles: true }));
+      batmanChildItem.shadowRoot.querySelector('button').dispatchEvent(new MouseEvent('click', { bubbles: true }));
+      batmanChildItem.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       await page.waitForChanges();
 
       expect(batmanItem.expanded).toBe(true);
@@ -180,8 +210,8 @@ describe('mg-menu', () => {
 
       // open joker item must close batman first child item
       const jokerItem: HTMLMgMenuItemElement = page.doc.querySelector('[title="joker"]').closest('mg-menu-item');
-
-      jokerItem.shadowRoot.querySelector('button').dispatchEvent(new CustomEvent(event, { bubbles: true }));
+      jokerItem.shadowRoot.querySelector('button').dispatchEvent(new MouseEvent(event, { bubbles: true }));
+      jokerItem.dispatchEvent(new MouseEvent(event, { bubbles: true }));
       await page.waitForChanges();
 
       expect(batmanItem.expanded).toBe(false);
@@ -197,7 +227,7 @@ describe('mg-menu', () => {
       const mutations = [[{ type: 'childList', removedNodes: [{ nodeName: 'MG-ITEM-MORE' }] }], [{ type: 'childList', removedNodes: [{ nodeName: 'MG-MENU-ITEM' }] }]];
 
       for await (const mutation of mutations) {
-        fireMo(mutation);
+        fireMo[0](mutation);
 
         await page.waitForChanges();
 
@@ -207,8 +237,18 @@ describe('mg-menu', () => {
   });
 
   describe.each([true, false])('mg-item-more, with submenu %s', submenu => {
+    test('should update "itemmore" children direction', async () => {
+      const page = await getPage({ label: 'batman menu' }, { submenu, itemMore: true, siblingMenu: false });
+
+      page.doc.querySelector('mg-item-more').shadowRoot.querySelector('mg-menu').appendChild(page.doc.querySelector('mg-menu-item'))
+      fireMo[0]({type: 'childList'})
+      await page.waitForChanges();
+
+      expect(page.root).toMatchSnapshot();
+    });
+
     test('should manage "itemmore" prop update', async () => {
-      const page = await getPage({ label: 'batman menu' }, { submenu, itemMore: true });
+      const page = await getPage({ label: 'batman menu' }, { submenu, itemMore: true, siblingMenu: false });
 
       page.doc.querySelector('mg-menu').itemmore = { icon: { icon: 'user' } };
 
@@ -218,7 +258,7 @@ describe('mg-menu', () => {
     });
 
     test.each(['click', 'focus'])('should manage "mg-item-more" event %s', async event => {
-      const page = await getPage({ label: 'batman menu' }, { submenu, itemMore: true });
+      const page = await getPage({ label: 'batman menu' }, { submenu, itemMore: true, siblingMenu: false });
 
       const mgItemMore = page.doc.querySelector('mg-item-more');
       const moreMenuItem = mgItemMore.shadowRoot.querySelector('mg-menu-item');