+++ packages/linting-styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mgdis/linting-styles
 
+## 1.0.1
+
+### Patch Changes
+
+- 2d90802: Update dependencies
+
 ## 1.0.0
 
 ### Major Changes