+++ packages/mg-components-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 4.1.12
+
+### Patch Changes
+
+- 2d90802: Update dependencies
+
 ## 4.1.11
 
 ### Patch Changes